requirement.txt
Django==1.11
MySQL-python==1.2.5
Pillow==4.0.0
django-simple-captcha==0.5.5
django-pure-pagination==0.3.0
pycrypto==2.6.1
ansible==2.3.2.0
passlib==1.7.1
paramiko==2.2.1
termcolor==1.1.0
prompt-toolkit==1.0.15
pygments==2.2.0
pyte==0.6.0
